EstatisticasService finalizado

+++ src/main/java/com/danilo/transacao_api/controller/dtos/EstatisticasResponseDTO.java
@@ -1,4 +1,4 @@
 package com.danilo.transacao_api.controller.dtos;
 
-public record EstatisticasResponseDTO(Integer count, Double sum, Double avg, Double min, Double max) {
+public record EstatisticasResponseDTO(Long count, Double sum, Double avg, Double min, Double max) {
 }
